pyproject.toml
[tool.poetry]
name = "py-aiger-discrete"
version = "0.1.0"
description = "Library for modeling functions over discrete sets using aiger circuits."
authors = ["Marcell Vazquez-Chanlatte <mvc@linux.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
py-aiger-bv = "^4.3.0"
funcy = "^1.14"
pyrsistent = "^0.17.3"
py-aiger-ptltl = "^3.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
pytest-flake8 = "^1.0.6"
hypothesis = "^5.30.0"
pytest-cov = "^2.10.1"
codecov = "^2.1.9"
pdoc3 = "^0.9.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
